prctl: adds PR_SET_TIMERSLACK_PID for setting timer slack of an arbitrary thread.
Second argument is similar to PR_SET_TIMERSLACK, if non-zero then the
slack is set to that value otherwise sets it to the default for the thread.

Takes PID of the thread as the third argument.

This allows power/performance management software to set timer slack for
other threads according to its policy for the thread (such as when the
thread is designated foreground vs. background activity)

Change-Id: I81af1c2f5a97b9e4273da16fa8a0628f14364a76
Signed-off-by: Ruchi Kandoi <kandoiruchi@google.com>

Conflicts:
	kernel/sys.c